Shooter fixes & deprecations
Deprecated old/useless shooter classes and also finished some of the TODO comments in shooter. One comment said remove follower bindings but I didn't see any so not sure on that.

+++ ECU/src/main/java/frc/robot/OI.java
@@ -65,18 +65,18 @@ public class OI {
     /**
      * Returns state of wrist up binding as a double
      *
-     * @return double 1 if wrist up is pressed, else 0
+     * @return int 1 if wrist up is pressed, else 0
      */
-    public double getWristUp() {
+    public int getWristUp() {
         return controller.getRawAxis(OIConstants.WRIST_UP_AXIS) > 0.5 ? 1 : 0;
     }
 
     /**
      * Returns state of wrist down binding as a double
      *
-     * @return double 1 if wrist down is pressed, else 0
+     * @return int 1 if wrist down is pressed, else 0
      */
-    public double getWristDown() {
+    public int getWristDown() {
         return controller.getRawAxis(OIConstants.WRIST_DOWN_AXIS) > 0.5 ? 1 : 0;
     }
